feat(connector): handle single and bulk connectors with the same account number fix(patch): update connector settings
