+++ india_banking_connector/connectors/bank_connector.py
@@ -10,6 +10,7 @@ from Crypto.PublicKey import RSA
 from Crypto.Util.Padding import pad, unpad
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
+from frappe import _
 from frappe.model.document import Document
 from frappe.query_builder import DocType
 from jose import jwe, jws
@@ -19,6 +20,13 @@ class BankConnector(Document):
 	def __init__(self, *args, **kwargs):
 		super().__init__(*args, **kwargs)
 		self.validate_user_permission()
+
+	def get_response(self, method):
+		self.is_active()
+		if method and hasattr(self, method):
+			return getattr(self, method)()
+		else:
+			frappe.throw(_("Invalid Method"))
 
 	def is_active(self):
 		if not self.active:
