+++ india_banking_connector/patches.txt
@@ -6,3 +6,4 @@ india_banking_connector.patches.v1_0.create_bank_doctype # apr29
 india_banking_connector.patches.v1_0.update_connector_settings #sep23
 india_banking_connector.patches.v1_0.update_new_api_endpoint #apr29
 india_banking_connector.patches.v1_0.update_standard_banks #sep28
+india_banking_connector.patches.v15_0.update_connector_settings #Oct31
