+++ india_banking_connector/patches/v15_0/update_connector_settings.py
@@ -1,0 +1,10 @@
+import click
+import frappe
+
+from india_banking_connector.install import create_connector_settings
+
+
+def execute():
+	click.echo("* Removing Connector Map... *")
+	frappe.db.delete("Connector Map")
+	create_connector_settings()
